+++ sales-coach-avatar/src/providers/index.ts
